Combine SuiteRunner with AbstractRunner
