+++ src/main/scala/scala/tools/partest/nest/AbstractRunner.scala
@@ -10,8 +10,11 @@ package nest
 import utils.Properties._
 import scala.tools.nsc.Properties.{ versionMsg, propOrFalse, setProp }
 import scala.collection.{ mutable, immutable }
+import scala.util.{ Try, Success, Failure }
 import TestKinds._
 import scala.reflect.internal.util.Collections.distinctBy
+import java.util.concurrent.Executors
+import java.util.concurrent.TimeUnit.NANOSECONDS
 
 abstract class AbstractRunner {
 
@@ -25,8 +28,6 @@ abstract class AbstractRunner {
     logOnFail = config.optShowLog,
     colorEnabled = colorEnabled
   )
-
-  val suiteRunner: SuiteRunner
 
   protected val printSummary         = true
   protected val partestCmd           = "test/partest"
@@ -116,7 +117,7 @@ abstract class AbstractRunner {
       config.optTimeout foreach (x => setProp("partest.timeout", x))
 
       if (!nestUI.terse)
-        nestUI.echo(suiteRunner.banner)
+        nestUI.echo(banner)
 
       val partestTests = (
         if (config.optSelfTest) TestKinds.testsForPartest
@@ -173,7 +174,7 @@ abstract class AbstractRunner {
           val num = paths.size
           val ss = if (num == 1) "" else "s"
           comment(s"starting $num test$ss in $kind")
-          val results = suiteRunner.runTestsForFiles(paths map (_.jfile.getAbsoluteFile), kind)
+          val results = runTestsForFiles(paths map (_.jfile.getAbsoluteFile), kind)
           val (passed, failed) = results partition (_.isOk)
 
           passedTests ++= passed
@@ -188,5 +189,115 @@ abstract class AbstractRunner {
       issueSummaryReport()
     }
     isSuccess
+  }
+
+  //---- from SuiteRunner:
+
+  val fileManager: FileManager
+
+  val updateCheck: Boolean = config.optUpdateCheck
+  val failed: Boolean = config.optFailed
+  val javaCmdPath: String = PartestDefaults.javaCmd
+  val javacCmdPath: String = PartestDefaults.javacCmd
+  val scalacExtraArgs: Seq[String] = Seq.empty
+  val javaOpts: String = PartestDefaults.javaOpts
+  val scalacOpts: String = PartestDefaults.scalacOpts
+
+  import PartestDefaults.{ numThreads, waitTime }
+
+  setUncaughtHandler
+
+  def banner = {
+    val baseDir = fileManager.compilerUnderTest.parent.toString
+    def relativize(path: String) = path.replace(baseDir, s"$$baseDir").replace(PathSettings.srcDir.toString, "$sourceDir")
+    val vmBin  = javaHome + fileSeparator + "bin"
+    val vmName = "%s (build %s, %s)".format(javaVmName, javaVmVersion, javaVmInfo)
+
+  s"""|Partest version:     ${Properties.versionNumberString}
+      |Compiler under test: ${relativize(fileManager.compilerUnderTest.getAbsolutePath)}
+      |Scala version is:    $versionMsg
+      |Scalac options are:  ${(scalacExtraArgs ++ scalacOpts.split(' ')).mkString(" ")}
+      |Compilation Path:    ${relativize(FileManager.joinPaths(fileManager.testClassPath))}
+      |Java binaries in:    $vmBin
+      |Java runtime is:     $vmName
+      |Java options are:    $javaOpts
+      |baseDir:             $baseDir
+      |sourceDir:           ${PathSettings.srcDir}
+    """.stripMargin
+    // |Available processors:       ${Runtime.getRuntime().availableProcessors()}
+    // |Java Classpath:             ${sys.props("java.class.path")}
+  }
+
+  def onFinishTest(testFile: File, result: TestState): TestState = result
+
+  def runTest(testFile: File): TestState = {
+    val runner = new Runner(testFile, this, nestUI)
+
+    // when option "--failed" is provided execute test only if log
+    // is present (which means it failed before)
+    val state =
+      if (failed && !runner.logFile.canRead)
+        runner.genPass()
+      else {
+        val (state, _) =
+          try timed(runner.run())
+          catch {
+            case t: Throwable => throw new RuntimeException(s"Error running $testFile", t)
+          }
+        nestUI.reportTest(state, runner)
+        runner.cleanup()
+        state
+      }
+    onFinishTest(testFile, state)
+  }
+
+  def runTestsForFiles(kindFiles: Array[File], kind: String): Array[TestState] = {
+    nestUI.resetTestNumber(kindFiles.size)
+
+    val pool              = Executors.newFixedThreadPool(numThreads)
+    val futures           = kindFiles map (f => pool submit callable(runTest(f.getAbsoluteFile)))
+
+    pool.shutdown()
+    Try (pool.awaitTermination(waitTime) {
+      throw TimeoutException(waitTime)
+    }) match {
+      case Success(_) => futures map (_.get)
+      case Failure(e) =>
+        e match {
+          case TimeoutException(d)      =>
+            nestUI.warning("Thread pool timeout elapsed before all tests were complete!")
+          case ie: InterruptedException =>
+            nestUI.warning("Thread pool was interrupted")
+            ie.printStackTrace()
+        }
+        pool.shutdownNow()     // little point in continuing
+        // try to get as many completions as possible, in case someone cares
+        val results = for (f <- futures) yield {
+          try {
+            Some(f.get(0, NANOSECONDS))
+          } catch {
+            case _: Throwable => None
+          }
+        }
+        results.flatten
+    }
+  }
+
+  class TestTranscript {
+    private val buf = mutable.ListBuffer[String]()
+
+    def add(action: String): this.type = { buf += action ; this }
+    def append(text: String) { val s = buf.last ; buf.trimEnd(1) ; buf += (s + text) }
+
+    // Colorize prompts according to pass/fail
+    def fail: List[String] = {
+      import nestUI.color._
+      def pass(s: String) = bold(green("% ")) + s
+      def fail(s: String) = bold(red("% ")) + s
+      buf.toList match {
+        case Nil  => Nil
+        case xs   => (xs.init map pass) :+ fail(xs.last)
+      }
+    }
   }
 }
